+++ src/main/java/com/textbasedgame/items/genericItems/keyItem.java
@@ -1,0 +1,5 @@
+package com.textbasedgame.items.genericItems;
+
+public abstract class keyItem {
+    public abstract void Use();
+}