+++ src/main/java/com/textbasedgame/items/keyItems/chestKey.java
@@ -1,0 +1,11 @@
+package com.textbasedgame.items.keyItems;
+
+import com.textbasedgame.items.genericItems.keyItem;
+import com.textbasedgame.playerFiles.player;
+
+public class chestKey extends keyItem{
+    @Override
+    public void Use(){
+        player.keyItemInventory.remove(this);
+    }
+}